fix -Wsign-compare warnings
"comparison between signed and unsigned integer expressions"

+++ src/lib.c
@@ -193,7 +193,7 @@ int dav1d_decode(Dav1dContext *const c,
         if ((res = dav1d_parse_obus(c, in)) < 0)
             return res;
 
-        assert(res <= in->sz);
+        assert(res <= (int)in->sz);
         in->sz -= res;
         in->data += res;
         if (!in->sz) dav1d_data_unref(in);
@@ -252,7 +252,7 @@ void dav1d_close(Dav1dContext **const c_out) {
                 t->tile_thread.die = 1;
             }
             pthread_cond_broadcast(&f->tile_thread.cond);
-            while (f->tile_thread.available != (1 << f->n_tc) - 1)
+            while (f->tile_thread.available != (1U << f->n_tc) - 1)
                 pthread_cond_wait(&f->tile_thread.icond,
                                   &f->tile_thread.lock);
             pthread_mutex_unlock(&f->tile_thread.lock);
